Remove Filter from Chain component
