+++ packages/frontend/src/components/Chain/index.ts
@@ -1,3 +1,2 @@
 export * from './Chain';
 export * from './Tab';
-export * from './Filter';
